feature #32: use currentUser in maps microservice

+++ maps/src/main/java/com/mypin/maps/services/MapsService.java
@@ -10,6 +10,8 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.mypin.maps.clients.INotificationsFeignClient;
@@ -21,6 +23,7 @@ import com.mypin.maps.dtos.SharingDto;
 import com.mypin.maps.dtos.SynchronizationDto;
 import com.mypin.maps.enums.MapSort;
 import com.mypin.maps.enums.NotificationType;
+import com.mypin.maps.exceptions.ForbiddenException;
 import com.mypin.maps.exceptions.ResourceNotFoundException;
 import com.mypin.maps.models.Map;
 import com.mypin.maps.models.Sharing;
@@ -39,7 +42,8 @@ public class MapsService implements IMapsService {
 	private final ISynchronizationFeignClient synchronizationFeignClient;
 
 	public MapsService(IMapsRepository repository, ISharingRepository sharingRepository,
-			IUsersFeignClient usersFeignClient, INotificationsFeignClient notificationsFeignClient, ISynchronizationFeignClient synchronizationFeignClient) {
+			IUsersFeignClient usersFeignClient, INotificationsFeignClient notificationsFeignClient,
+			ISynchronizationFeignClient synchronizationFeignClient) {
 		super();
 		this.repository = repository;
 		this.sharingRepository = sharingRepository;
@@ -52,18 +56,21 @@ public class MapsService implements IMapsService {
 	public Map save(Map map) {
 		map.setCreatedDate(new Date());
 		map.setUpdatedDate(new Date());
-		// TODO: save user id
+		map.setOwnerId(getCurrentUser());
 		return repository.save(map);
 	}
 
 	@Override
 	public Map update(Map updated, UUID id) {
-		// TODO: check if right user
+
 		Optional<Map> map = repository.findById(id);
 		if (map.isEmpty()) {
 			throw new ResourceNotFoundException("Map not found");
 		}
 		Map existingMap = map.get();
+		if (!hasPermissions(existingMap)) {
+			throw new ForbiddenException("User does not have rights to map");
+		}
 		existingMap.setTitle(updated.getTitle());
 		existingMap.setUpdatedDate(new Date());
 		return repository.save(existingMap);
@@ -80,22 +87,27 @@ public class MapsService implements IMapsService {
 
 	@Override
 	public void delete(UUID id) {
-		// TODO: check if right user
 		Optional<Map> map = repository.findById(id);
 		if (map.isEmpty()) {
 			throw new ResourceNotFoundException("Map not found");
+		}
+		if (!hasFullPermissions(map.get())) {
+			throw new ForbiddenException("User does not have rights to map");
 		}
 		repository.deleteById(id);
 	}
 
 	@Override
 	public Map patchTitle(String title, UUID id) {
-		// TODO: check if right user
 		Optional<Map> map = repository.findById(id);
 		if (map.isEmpty()) {
 			throw new ResourceNotFoundException("Map not found");
 		}
 		Map existingMap = map.get();
+		if (!hasPermissions(existingMap)) {
+			throw new ForbiddenException("User does not have rights to map");
+		}
+
 		existingMap.setTitle(title);
 		existingMap.setUpdatedDate(new Date());
 		synchronizeMap(existingMap);
@@ -109,6 +121,9 @@ public class MapsService implements IMapsService {
 			throw new ResourceNotFoundException("Map not found");
 		}
 		Map existingMap = map.get();
+		if (!hasPermissions(existingMap)) {
+			throw new ForbiddenException("User does not have rights to map");
+		}
 		existingMap.setUpdatedDate(new Date());
 		synchronizeMap(existingMap);
 		return repository.save(existingMap);
@@ -120,12 +135,12 @@ public class MapsService implements IMapsService {
 		if (map.isEmpty()) {
 			throw new ResourceNotFoundException("Map not found");
 		}
-		
+
 		AppUserDto user = usersFeignClient.getByEmail(sharingDto.email).getBody();
 		if (user == null) {
 			throw new ResourceNotFoundException("User with email not found");
 		}
-		
+
 		Sharing sharing = new Sharing();
 		sharing.setMapId(id);
 		sharing.setUserId(user.id);
@@ -139,7 +154,7 @@ public class MapsService implements IMapsService {
 		Map map = get(id);
 		List<Sharing> sharingList = sharingRepository.getByMapId(id);
 		if (!sharingList.isEmpty()) {
-			final UUID currentUserId = UUID.randomUUID();// TODO: get current userId
+			final UUID currentUserId = getCurrentUser();
 			List<NotificationDto> notifications = new ArrayList<>();
 			notifications.addAll(sharingList.stream().map(fillCallToMapNotification(currentUserId)).toList());
 			notifications.add(fillCallToMapNotification(currentUserId, map.getOwnerId()));
@@ -158,29 +173,49 @@ public class MapsService implements IMapsService {
 	private Function<? super Sharing, NotificationDto> fillCallToMapNotification(UUID currentUserId) {
 		return sharing -> fillCallToMapNotification(sharing.getUserId(), currentUserId);
 	}
-	
+
 	private NotificationDto fillCallToMapNotification(UUID callingUserId, UUID targetUserId) {
 		return new NotificationDto(targetUserId, callingUserId, NotificationType.CALL_TO_MAP, "call");
 	}
 
 	@Override
 	public List<Map> search(String title, MapSort sort, Boolean isSharedWithMe, Boolean isMyOwn) {
-		//TODO: get user id
-		Specification<Map> specification = MapSpecification.filter(title, sort, isMyOwn, isSharedWithMe, UUID.fromString("accd2a12-8d3b-438e-8a73-d12e3419030b"));
+		UUID user = getCurrentUser();
+		Specification<Map> specification = MapSpecification.filter(title, sort, isMyOwn, isSharedWithMe, user);
 		return repository.findAll(specification);
-
 	}
-	
+
 	private void synchronizeMap(Map map) {
 		SynchronizationDto synchronizationDto = new SynchronizationDto();
 		synchronizationDto.channel = map.getId().toString();
 		synchronizationDto.content = ""; // TODO fill map
 		synchronizationFeignClient.sendSynchronizationMessage(synchronizationDto);
 	}
-	
+
 	private void sendSharingNotification(Sharing sharing) {
-		NotificationDto notification = new NotificationDto(sharing.getUserId(), sharing.getUserId(), NotificationType.MAP_SHARED_TO_USER, "MAP_SHARED_TO_USER"); //TODO: fill content
+		NotificationDto notification = new NotificationDto(sharing.getUserId(), sharing.getUserId(),
+				NotificationType.MAP_SHARED_TO_USER, "MAP_SHARED_TO_USER"); // TODO: fill content
 		notificationsFeignClient.create(notification);
+	}
+
+	private UUID getCurrentUser() {
+		SecurityContext context = SecurityContextHolder.getContext();
+		String uuid = context.getAuthentication().getName();
+
+		System.out.println("name: " + uuid);
+		return UUID.fromString(uuid);
+	}
+
+	private boolean hasFullPermissions(Map map) {
+		UUID user = getCurrentUser();
+		return map.getOwnerId().equals(user);
+	}
+
+	private boolean hasPermissions(Map map) {
+		UUID user = getCurrentUser();
+		List<Sharing> sharingList = sharingRepository.getByMapId(map.getId());
+		List<UUID> usersWithPermissions = sharingList.stream().map(sharing -> sharing.getUserId()).toList();
+		return usersWithPermissions.contains(user);
 	}
 
 }
